[TASK] Streamline functional test snapshots
The testing-framework functional test db snapshot
helper expects the database to be empty when
withDatabaseSnapshot() is called. Rows should
only be populated within the create callback.

There was a workaround in testing-framework to
relax this restriction due to unfortunate backend
user related setup code flow. This has been resolved
with #97928, so the testing-framework snapshot can
be more strict on this now.

The patch adapts a couple of snapshot usages to
align to this enforced restriction.

$ composer req --dev typo3/testing-framework:dev-main

Change-Id: I50d6dda3671729bd603899e37e7c7d0979c094fe
Resolves: #97935
Related: #97928
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75174
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>