[!!!][TASK] Remove border and longdescURL properties from IMAGE cObject
HTML deprecated the "border" attribute and the "longdesc"
attributes which are not set anymore by TYPO3's Frontend.

This mainly relates to "IMAGE" ContentObject.

TypoScript properties
* IMAGE.border
* IMAGE.longdescURL
are no longer used.

The helper function "imgTag" in GraphicalFunctions
is also removed.

Fluid's Thumbnail ViewHelper in Backend
does not allow "longdesc" anymore. As this ViewHelper is
internal, it does not affect public API.

References:
* https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/border
* https://developer.mozilla.org/en-US/docs/Web/API/HTMLImageElement/longDesc

Resolves: #98308
Releases: main
Change-Id: I5396c6b01a2157b132278ce388150b793f2d3bbd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75602
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>