[BUGFIX] List invalid field in FormEngine review
The FormEngine review module broke with the introduction of Bootstrap 5
due to changed API of its Popover module.

Our internal Popover wrapper is now used to render the list of failed
fields again. This also introduces another bugfix in the Popover wrapper
which updates the Popover's content once it was rendered already.

Additionally, a nasty workaround is added: previously, the field id was
stored as a data attribute to the list item in the validation list.
Somewhere in the Popover > Bootstrap > PopperJS handling, those custom
data attributes are removed before rendering the HTML. To circumvent
this issue, the field is now added as last element in the `classList`
DOMTokenList for later use.

Resolves: #94774
Releases: main, 11.5
Change-Id: I00d8196102f08548c8ec3433c1d0f0c8c7f05d04
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75612
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>