[BUGFIX] Correct Backend Scaffold
During the migration to bootstrap the scaffold was reworked and is now flowing
incorrectly. The introduction flex on the body has multiple side effects, for example
incorrect layering of modals that are anchored directly on the body.

This patch aims to reduce the differences on different implementations of the
scaffold, as well as removing the flex bugs.

Resolves: #98263
Releases: main
Change-Id: I900c44d56243aa0cff4b4f6561f456991a4d7295
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75614
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>