[BUGFIX] Properly evaluate whether toolbar item provides a dropdown
The ToolbarItemInterface requires the methods
hasDropDown() as well as getDropDown(). Since
#98214 the evaluation whether a toolbar item
provides a dropdown is done in Fluid, using the
corresponding magic "{toolbarItem.dropDown}".

Fluid however first checks for a "Getter" and
therefore uses the getDropDown() method instead
of the hasDropDown() method to determine whether
a toolbar item provides a dropdown. This led to
display errors in case a toolbar item only provides
a dropdown for some use cases as this was now
always used.

This is now fixed using a dedicated conditional
ViewHelper.

Resolves: #98299
Related: #98214
Releases: main
Change-Id: I3aabfd0debe95ae35f2d9459c483d55e2260292b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75663
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>