[TASK] Streamline colors, correct contrast colors
The core currently uses a wide range of
state colors across the backend. This is
a mixture of defined colors for specific areas
and the global usage of the Bootstrap 5
theme colors.

The current state is a mixture of vibrant
and faded colors. The current implementation
results in false calculated contrast colors for
certain areas.

To resolve this we are now unifying all used
colors across the backend and streamline
the text color contrast calculation.

That needed several cleanups across the
backend regarding markup and CSS styles.
The result is a more harmonized look and
feel of the Backend.

Affected areas are:
- Alerts
- Badges
- Buttons
- Callouts
- Panels
- Modals
- Tables

Resolves: #98349
Releases: main

Change-Id: I4311e35bc9e1dcebb8cca9e2af26c86735656155
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75692
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>