[TASK] Streamline FormEngine inline-relation JavaScript module loading
Support for non JavaScriptModuleInstruction RequireJS invocations
in FormEngine has been dropped in #96221, but the special ajax
inline-relation was kept by accident. This is now streamlined to
throw the same exception that FormResultCompiler throws for non
JavaScriptModuleInstructions.

Also the internal FormResultTrait is removed as
serializeJavaScriptModuleInstructionItems() is no longer used.
(Note that is was already unused, as the only remaing user
createExecutableStringRepresentationOfRegisteredRequireJsModules()
was only called with skipInstructions=true which skipped the
serialization in favor of direct JavaScriptItem passing).

Releases: main
Resolves: #98422
Related: #96221
Change-Id: I5b5c8a36dca8f74a4c498a786a35d73e2ac4b343
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75853
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>