[!!!][TASK] Remove jQuery from FormEngine auto-suggest
This patch migrates the auto-suggest feature used in FormEngine to a
custom implementation incorporating web components.
A user may navigate through the result list with the keyboard's arrow
keys and select an item with the "Enter" key.

In the same run, the JavaScript library `devbridge-autocomplete` is
removed as it is not used anymore within TYPO3, along with its
associated CSS definitions.

As the web component approach allows us more flexibility and to maintain
a consistent look & feel of the backend, some properties were removed
from the suggest wizard's items.

Resolves: #98455
Releases: main
Change-Id: I7abc40bfe9161a7a246ac451bd046034dcc8d9bd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75881
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>