[TASK] Do not use jQuery and jQuery UI in page module drag&drop
This patch reimplements the drag&drop functionality in the page module
with interactjs and thus removes the need of jQuery and jQuery UI, which
is removed.

Due to the fact the iframe is not scrollable, but the module body is
scrollable, we need to intercept the position calculation at some
places.

Resolves: #98471
Releases: main
Change-Id: I0ea04795e3e03cd35487ce828a6887b4cdaf181d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75887
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>