[TASK] Replace `imagesloaded` with native browser API
The library `imagesloaded` is only used in the image cropper to
initialize the cropper functionality once the image is loaded. This can
be achieved with native browser API, `imagesloaded` is therefor not
needed anymore and thus removed.

See https://caniuse.com/mdn-api_htmlimageelement_complete
for compatibility.

Executed command:

    npm remove --save imagesloaded @types/imagesloaded

Resolves: #98548
Releases: main
Change-Id: I154ef22b41febb3788599e8220dd55f6989c3048
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76001
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>