[BUGFIX] Animate docheader folding via CSS transitions
A backend module's docheader is "folded" when scrolling down in the
backend module. Sometimes, the folding process appears to be laggy due
to the pixel-snapping approach of animating a position change, which also
leads to more work for browser engines.

For a better rendering performance, the element isn't physically moved in
the DOM, but transformed instead, offloading work to the GPU.

See [1] for further information.

[1] https://web.dev/animations-overview/

Resolves: #98693
Releases: main, 11.5
Change-Id: I93be697d8105eef83ce2376c8879419058dfdb85
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76218
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>