[BUGFIX] Make CE actions clickable after visibility toggle
When a content element gets hidden or visible, the page module is
reloaded with an anchor in the URL to scroll to the content element
that was updated.

After #98471 it was not possible anymore to click the "+ Content"
button above such element unless it was dragged and dropped again into
the same column. The reason is that a pseudo element `:target:before`
is generated with a specific height and margin to overcome the height of
the floating module docheader, overlaying the CE action button.
In the patch mentioned before, relative positioning was added to the
columns as this is needed for the dropzones, finally causing this bug.

To solve this issue, the CE actions now have a higher z-index to be
clickable again.

Resolves: #98703
Related: #98471
Releases: main
Change-Id: I32d93a79842cb599526eae9dec3373c3d3fc31eb
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76234
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>