[TASK] Allow multiple actions and custom invoke handlers for search results
Search results now may have multiple actions to trigger, depending on
their context. For pages it'd make sense to either open them in the
backend or the frontend. Those actions are rendered next to the focussed
search result.

Invoke handlers may now get registered for custom search result items
that allow to modify the behavior when selecting a result item, e.g.
calculating and opening a different URL based on data that is not
available via PHP.

As a base example, the `PageRecordProvider` brings a custom invoke
handler for an action to open a page in a new browser tab.

This feature is purposely not documented as such as behavior or API may
change in further iterations.

Resolves: #99090
Releases: main
Change-Id: I71f0aa7b93359398b5a8011361dac5018c9f1cde
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76405
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>