[TASK] Deprecate old TypoScriptParser
The main goal of this patch is to avoid last usages
of old TypoScriptParser and to substitute it with
usages of the new parser approach.

A couple of side usages have to be adapted, but the
main work is to establish new factories for UserTsConfig
and PageTsConfig while deprecating the former approach
and adding functional test coverage to the factories.

BackendUtility::getPagesTSconfig() is still the main
API to retrieve PageTsConfig, for UserTsConfig it is
$backendUser->getTSConfig(). The new API is thus
considered @internal, at least for now.

Resolves: #99120
Related: #97816
Releases: main
Change-Id: I985929e1f2a90560059efe1e0f40f8c590ad1e4a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76565
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>