[FEATURE] Allow static backdrops for modals
The Modal API is now able to render a static backdrop to avoid closing
the modal when clicking it which may come in handy in case closing the
modal would result in a negative user experience, e.g. in the image
cropper.

The new boolean configuration option `staticBackdrop` controls whether
a static backdrop should be rendered or not, defaulting to `false`.

The trigger class "t3js-modal-trigger" also supports the new option
by setting corresponding data attribute.

Resolves: #99092
Releases: main
Change-Id: I0e1c7753944b41d5266aa6232d1fa838752f95b3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76605
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>