[FEATURE] Introduce TCA type "uuid"
A new TCA type "uuid" is introduced, which
allows to simplify the configuration when
working with fields, containing a UUID.

The functionality is based on the Uid
symfony component.

FormEngine will automatically create a UUID when
non is defined yet. Same does the DataHandler in
case an invalid UUID is defined, while the field
is defined as "required", which is the default.

The UUID is displayed as a readonly input field.

Using the new TCA type, corresponding database
columns are added automatically.

Resolves: #100171
Releases: main
Change-Id: Ic81d4cc0158e77988e38cdab6ddcc5d42aa47fcd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76891
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>