[BUGFIX] Fix exception with PHP 8.1 in ItemProcessingService
The pageId is already passed to the functon
ItemProcessingService::getProcessingItems(). This is now used instead
of retrieving the uid / pid from the row array.

The previous behaviour could result in an exception with PHP 8.1
since the uid in the row might not be initialized if a new page
is created. If a field with itemProcFunc existed for a field
in the pages table.

The page id which was passed to the function already handled
the case of empty uid by passing the uid of the parent page.

Resolves: #99295
Releases: main, 11.5
Change-Id: I9c72189c6f80e0a45cb9abfcc80541b091c126b8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76943
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>