[BUGFIX] Always display search box in file browser
The search box in the file browser is now displayed again,
even if no files are in the folder. This is necessary because
the search works recursively.

Actually, the search box was only displayed for an active search
with an empty result list, due to another bug. FileSearchResultInterface
is never "empty" since it's an object. This is now fixed as well.

Additionally, the flash message is improved, in case no files
are available for a search term.

Resolves: #99447
Releases: main, 11.5
Change-Id: I970b19c7779075711f97987e80aad3f376cd5f40
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77247
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>