[BUGFIX] Fix accordion visibility in image processing environment test
The display flex on the .modal-body was used to center the
loading animation in the modal screen but led to making
.panel-groups inaccessible. To avoid side effects we are
resetting the custom styles on .modal-body and defining the
height of .modal-loading to 100% of its parent. Since
modal-body also defines already whitespace, we are also
removing the top and bottom the margin on the first and last
element in that container, to provide a more evenly spaced
visual experience.

Releases: main
Resolves: #99590
Related: #98349
Change-Id: If540659cffa886c4b66da7cf6622b5a051eb4bce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77461
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>