[TASK] Add search field to active 'Page TSconfig'
The "Page TSconfig" backend module now has a search that
allows an integrator to search PageTS either by a settings
identifier (e.g. "mod.linkval") or a value (e.g. "textpic").

Partial searches are also supported, "wizardItems" resolves
to "mod.wizards.newContentElement.wizardItems".

The search value is stored in the browser session and retained
when the module is visited again. To avoid conflicts with the
internal `collapse-state-persister` module, a new setting is
introduced to disable the modules recovery behavior if a
specific key is available in localStorage or sessionStorage.

Resolves: #99602
Releases: main
Change-Id: If1b91d936597406b3d553e81b2ced1076bc054cf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77481
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>