[TASK] Improve "Active Typoscript" view
The module receives a similar treatment as
"Active PageTsConfig", namely client side
search and client side expand state persisting.

This simplifies server side a bit, which is
cleaned up. We're also avoiding the naming
'object browser' in file names and other places
in favor of 'active'. A series of minor bug
fixes is applied along the way.

Resolves: #99782
Related: #99747
Releases: main
Change-Id: Ibaef3123b3e6df316ef5e633582077fd0e702be0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77650
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>