[TASK] Clean up various TypeScript code for SVG tree
This change reduces some redundancy in implementations
for drag+drop for pagetree + folder tree,
also adds some comments and adds
some helper methods to make the code
a tiny bit more understandable.

In addition, to reduce our dependencies to
d3-drag, the DragDropHandlers now receive
the node and the original MouseEvent,
as there is nothing else that the DragDrop
handlers need for our solutions.

Resolves: #99778
Releases: main
Change-Id: Ia45a3ac5256d22ab38cf8c5d70fc4b939a53085a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77654
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>