[BUGFIX] Disable drag&drop from files within EXT:filelist module
This patch introduces a flag that enables itself once dragging was
started within the document of the filelist module. If the flag is
enabled, the drag&drop upload feature is *disabled* to avoid dragging
an image within the filelist and having a fullsize dropzone rendered,
which in fact accepts the dragged thumbnail and uploads it in the same
directory.

However, this will not prevent receiving images from anywhere else of
the TYPO3 backend as those are technically "external" files.

Resolves: #99800
Releases: main
Change-Id: I9b6184088b57690a8b588e6ad3b3f7cd16d369d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77671
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>