[BUGFIX] Support deferred ConsumerScope in GlobalEventHandler
The GlobalEventHandler changes the location of the current (target)
window. The location is set to a given destination without any ability
to interrupt. The existing ConsumerScope API isn't used even if the
current window is the ContentContainer. This patch extends the handling
of a navigation action by checking its target window and using proper
ContentContainer API to set the location. Therefore registered listener
can deny a location change and trigger own functions.

Resolves: #95474
Resolves: #100114
Releases: main, 12.4, 11.5
Change-Id: If6a3b1cc546e1f2ff0b8e9fed3f83055f3f55e49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79922
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>