[TASK] Update display of system information dropdown
The heading in the dropdown content now matches the label already
shown as button to open the dropdown in small viewports.

With #99742 a heading was introduced to structure the messages listed
below. With this change this heading will be removed again.

In addition, the standard link markup is now also used in the
introductory text.

The badge contains the amount of error messages. Badge position is now
aligned horizontally in the middle of the row in small viewports.

Resolves: #99853
Related: #99742
Related: #99538
Releases: main
Change-Id: Iea5a88c6b088ce810fc073a886251e8bfd1c880b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77721
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>