[TASK] Remove "direction" from site configuration
Site Configuration of a language has a lot of option,
one of them being "direction", defining the direction
of the language for Frontend Output.

Guess what? There are exactly 15 languages in the world
which have a "right-to-left" direction, and the HTML
standard actually defines that "ltr" (left to right)
does not need to be set. So this check is now done
automatically.

The option is now removed from the GUI, however
previous settings still apply.

Resolves: #99916
Releases: main
Change-Id: Icc98e6514658ab6915be25f012b02d2c0b76b38d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77814
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>