[TASK] Deprecate PageRenderer->enableDebugMode()
PageRenderer->enableDebugMode() is only called in
backend by the core. The properties set by the
method do not have any effect in BE during further
processing.

The method is typically not called in Frontend,
any calls can be moved to the single setters
instead.

The patch deprecates PageRenderer->enableDebugMode().

Resolves: #99932
Releases: main
Change-Id: Iddef89c9e97f721a7b33ccc838d20f17b03471f4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77831
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>