[TASK] Refine LiveSearch form behavior
The behavior of the LiveSearch form used in the modal window is refined
in several ways.

The most important change is the usage of `SearchDemand` to pre-fill the
search form. This required to add a `form` tag and to rename the
`searchField` field to `query` to align with the `SearchDemand`
requirements. Also, the `q` argument is renamed to `query` as well.

Instead of directly triggering the search at several locations, a form
submit is request. This allows to call `search()` only once and to
provide a `FormData` object, containing all fields and values for the
underlying AJAX request. With this, there is no need to store the values
in properties anymore (`this.searchTerm` and `this.searchProperties`),
which are dropped now.

Resolves: #99954
Releases: main
Change-Id: Ibeb8c45048f8cc6c4801df648a19c0ecf7afb883
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77853
Tested-by: core-ci <typo3@b13.com>
Tested-by: Frank Nägler <frank.naegler@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.com>