[BUGFIX] Position sidebar in image cropping correctly on the right
Currently, the sidebar containing the cropVariants
is always positioned below.

This has been fixed by adding a missing display flex
to the modal on sm-viewport and up.

On top of that the undefined this.cropBox error
on modal load was fixed by adding a empty JQuery object
on initialization.

Resolves: #99953
Releases: main
Change-Id: I679b188e38add5e06ad8e6334ddc15de01a42e60
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77865
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>