[FEATURE] Enable list/tile view for resources in link browser
With this change, we are rolling out the universal file-list
rendering for files and folders to the link browser. The link
browser implementation for files and folders is now part of the
file list extension.

The link browser now allows the user to choose the display type
of resources to match the personal preference between list and
tile rendering.

When the user now edits a link for a folder, the entry point is
now the parent folder of the selected element folder instead of
showing the contents of the selected resource. The user now sees
the selected folder in the presented list, this behavior mimics
the handling of selected files.

Resolves: #100206
Releases: main
Change-Id: Ib06257b9215ab74cff5039defd56785e63ce0e86
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78154
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>