[TASK] Add pagination to LiveSearch
The `SearchRepository` class used in the LiveSearch is now able to
handle pagination, allowing to render a pagination in the search modal
window. The `SearchProviderInterface` received a new method `count()`
used to calculcate the amount of potential results.

Since pagination is available now, the "Show all" button in the
LiveSearch serves no real purpose anymore and is therefore removed.

Resolves: #100289
Releases: main
Change-Id: I2f1f431d8d385d0ba1963483c2dd34b1e44cca67
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78232
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>