[BUGFIX] Do not send table wizard fields in payload
The fields used in the table wizard are only used for internal handling
to be able to compose the final table markup sent to DataHandler. The
"management fields" are not interesting on server side and therefore
their respective `name` attributes are removed to avoid sending
potentially huge, superfluous data.

Resolves: #100354
Releases: main, 12.4
Change-Id: If833f2bd1208a050dc88bc1844a19d9fbb209614
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78330
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Torben Hansen <derhansen@gmail.com>