[BUGFIX] Sanitize BackendUtility::titleAttribForPages()
When the "Access" module is called without a selected
page, the root page icon rendering triggers a PHP 8.0
array access warning.

This was triggered by the refactoring with #100454, but
is a general issue of the method, which is why the patch
should go to 11.5 as well.

The patch sanitizes $row['uid']. Other parts of the
method access further array keys without fallback,
but those should be fine - at least for now - since
this is about 'pages', and if there is a uid, the
other fields are probably provided as well.

Resolves: #100527
Related: #100454
Releases: main, 11.5
Change-Id: I40df9fcb86a145cfa68f7dc3086649c52efe7d79
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78558
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>