[BUGFIX] Finetune table-fit border removals
To avoid double borders at the bottom for tables in
a .table-fit wrapper the default border-bottom, for
the last row is removed.

The current version forgot to consider that there can be
multiple tbody or tfoot. The selector is now adjusted to
consider this.

For bordered tables, a double border appeared at the top,
this one is now also correctly removed. Captions are now
supported in a .table-fit wrapper and placed correctly.

Resolves: #100533
Releases: main
Change-Id: I1ce699c175cb4248e4803b872355918e6f22cb24
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78543
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>