[TASK] Improve beuser templates
Templates were streamlined to provide the same experience
across the available modules. That means, record counter
headlines, and icons, are now placed the same way.

Template names were consolidated and are now sorting
according to their responsibility, regardless that everything
is pushed into the backend user controller.

BackendUser -> BackendUser
BackendUserGroup -> BackendUserGroup
Filemount -> Filemount

This structure was already present in the partials
folder and was adapted also for the templates.

Partials that would just move information around without
a greater benefit was removed and information was moved
to the main template. Bloating the amount partials only
makes it harder to get an overview of the current structure
and the template.

Resolves: #100536
Releases: main
Change-Id: I9f37831ff6ad4623151c162049bf2f1fdfd0a6a0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78545
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>
Reviewed-by: Jasmina Ließmann <minapokhalo+typo3@gmail.com>