[BUGFIX] Fix page reloading on click on preview button (again)
The changes introduced in #93706 broke the preview button functionality
in multidomain setups. When first previewing the page in one page tree
and then clicking the preview link in the page tree of a different
domain, a DOMException due to a cross origin error would be triggered
in JavaScript, because it is not allowed to access the constructor of
a window object from a different context.

The thrown DOMException is now properly caught.

Resolves: #99286
Related: #93706
Releases: main, 11.5
Change-Id: I709f06ef89e248d7b88d500281e708f56557a176
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78571
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>