[BUGFIX] Correct link in PageMovingPagePositionMap
Repositioning a content element to a different page
using the list module "Re-position content element"
view fails with a broken "backend-in-backend" link.

The patch avoids linkThisScript() by handing the request
around and creates the url using the UriBuilder instead.

Note the construct is one of the few left over controllers
that have not been refactored yet. This should happen in
v13, together with re-thinking the view along the way.

Resolves: #100564
Releases: main
Change-Id: Ib912e878647ea778ecc941a185f73d9e835059a1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78598
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>